CePNEM Fitting API

CePNEM.run_mcmc_10 — Function

run_mcmc_10(ys, v, θh, P; n_init=100000, n_iters=11000, lr_adjust=1.1, model=:nl10d)

Run MCMC chain for fitting the CePNEM model to the data.

Arguments:

ys: Vector of observed neural activity (GCaMP) data.

v: Vector of observed worm velocity data.

θh: Vector of observed worm head curvature data.

P: Vector of observed worm pumping (feeding) data.

n_init: Number of iterations for the initialization phase of the MCMC chain.

n_iters: Number of iterations for the MCMC chain.

lr_adjust: Learning rate adjustment factor.

model: Model to use for the MCMC chain. Can be either :nl10c or :nl10d. It is recommended to use :nl10d.

Returns:

traces_fit: Vector of traces (ie: fit model parameters) for the MCMC chain.

accept: Matrix of acceptance rates for the MCMC chain.

δ_vals: Vector of step sizes for the MCMC chain.

CePNEM.nl10d_traces_to_params — Function

nl10d_traces_to_params(traces_fit)

Extracts the free parameters from the traces of the MCMC chain for the nl10d model.

Arguments:

traces_fit: Vector of traces (ie: fit model parameters) for the MCMC chain.

Returns:

A vector of the free parameters extracted from the traces of the MCMC chain for the nl10d model.