Simulation-Based Calibration API

CePNEM.rank_test — Function

rank_test(ground_truth, inferred, n_bins; print_extremes=false, normalize=true)

This function computes the rank test for a given ground truth and inferred values. It returns the ranks for each bin and each column of the inferred matrix.

Arguments:

ground_truth: a matrix of size (nsamples, nfeatures) representing the ground truth values.

inferred: a tensor of size (nsamples, ninferred, n_features) representing the inferred values from the model fit.

n_bins: an integer representing the number of bins to use for the rank test.

print_extremes: a boolean indicating whether to print the ground truth values that are at the extremes of the inferred values.

normalize: a boolean indicating whether to normalize the ranks by the number of samples.

Returns:

ranks: a matrix of size (nbins, nfeatures) representing the ranks for each bin and each column of the inferred matrix.

CePNEM.χ2_uniformtest — Function

χ2_uniformtest(raw_ranks)

This function performs a chi-squared test for uniformity on the given ranks.

Arguments:

raw_ranks: a vector of raw ranks.

Returns:

The p-value of the chi-squared test.